// src/shared/api/generated/model/billingInterval.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Marz API
 * Backend de Marz. Fuente de verdad del contrato HTTP entre marz-api y marz-front.
Se consume con `oapi-codegen` (server) y `openapi-typescript` + `openapi-fetch` (client).

 * OpenAPI spec version: 0.1.0
 */

export type BillingInterval = typeof BillingInterval[keyof typeof BillingInterval];


export const BillingInterval = {
  month: 'month',
  year: 'year',
} as const;